suma testeos a SaveFile con limpieza de log para falsos positivos

+++ 04-multiplication/src/domain/uses-cases/save-file.use-case.test.ts
@@ -2,12 +2,24 @@ import { SaveFile } from './save-file.use-case'
 import fs from 'fs'
 
 describe('SaveCaseFile', () => {
+  const customOptions = {
+    fileContent: 'custom content',
+    fileDestination: 'custom-outputs/file-destination',
+    fileName: 'custom-table-name'
+  }
+
+  const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.log`
+
   beforeEach(() => {
     //! clean up
     // fs.rmSync('outputs', { recursive: true })
   })
   afterEach(() => {
-    fs.rmSync('outputs', { recursive: true })
+    const outputFolderExist = fs.existsSync('outputs')
+    if (outputFolderExist) fs.rmSync('outputs', { recursive: true })
+
+    const customOutputFolderExist = fs.existsSync(customOptions.fileDestination)
+    if (customOutputFolderExist) fs.rmSync(customOptions.fileDestination, { recursive: true })
   })
 
   test('should save file with default values', () => {
@@ -26,5 +38,19 @@ describe('SaveCaseFile', () => {
       encoding: 'utf-8'
     })
     expect(fileContent).toBe(options.fileContent)
+  })
+
+  test('should save file with custom values', () => {
+    const saveFile = new SaveFile()
+
+    const result = saveFile.execute(customOptions)
+    expect(result).toBeTruthy()
+    const fileExist = fs.existsSync(customFilePath)
+    expect(fileExist).toBeTruthy()
+
+    const fileContent = fs.readFileSync(customFilePath, {
+      encoding: 'utf-8'
+    })
+    expect(fileContent).toBe(customOptions.fileContent)
   })
 })
